refactor: set up the storage and encoding crates
