+++ crates/storey-encoding/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "storey-encoding"
+description = "Interfaces for storey encodings"
+version = "0.1.0"
+edition = "2021"
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+categories.workspace = true
+keywords.workspace = true
+
+[dependencies]
