+++ crates/storey-storage/src/lib.rs
@@ -1,0 +1,5 @@
+mod backend;
+mod storage;
+
+pub use backend::{StorageBackend, StorageBackendMut};
+pub use storage::{IterableStorage, RevIterableStorage, Storage, StorageMut};
