+++ crates/storey/src/containers/mod.rs
@@ -1,6 +1,7 @@
 pub mod column;
 mod item;
 mod map;
+mod test_mocks;
 
 use std::marker::PhantomData;
